♻️ refactor: 스테이터스 레벨, 이름 추출을 StatusChart 컴포넌트에서 하도록 수정
